ProjectApp/requirements.txt
ipykernel
pandas
numpy
matplotlib
seaborn
wordcloud
scikit-learn
spacy
python-dotenv
streamlit
google-generativeai
streamlit
plotly